Eliminate runtime dependency on junit
This patch eliminates `kivakit-core` and several other projects' direct, non-test dependency on junit,
which causes dependees to inherit that dependency.

Doing so requires that we

  * Move `CoreUnitTest` and friends out of `kivakit-core` into a new `kivakit-test-support` project
  * Do the same for `UnitTest` in `kivakit-resource` --> `kivakit-resource-test-support`
  * Move tests for those two projects into their own separate tests-only projects, so attempting to
    use the test-support classes does not create a cyclic dependency (this is a Maven limitation)
    * There are new `kivakit-core-tests` and `kivakit-resource-tests` project for those tests

I've used an at least somewhat familiar naming convention of `*-test-support` for _projects containing
classes that help write tests_ and `*-tests` for _tests of some project_ (think `PropertyChangeSupport`)
to keep the purpose of the projects unambiguous.